refactor: extract shared Restart setup confirmation with explicit loss warning
The confirmation now names exactly what a restart discards (dir
selection, chosen host, recorded scan/login/build progress) and what it
keeps (files on disk, except a never-built directory setup itself
materialized). Extracted so Settings can render the same flow.

Part of docs/2026-07-08-onboarding-state-ownership.md (D6).
