+++ apps/native/src/components/widget/onboarding/onboarding-header.tsx
@@ -1,29 +1,9 @@
-import { ConfirmationDialog } from "@/components/widget/controls/confirmation-dialog";
-import { getTelemetry } from "@/lib/telemetry/instance";
-import { client } from "@/lib/orpc";
-import { onboardingActions } from "@nixmac/state";
+import { RestartSetupConfirmation } from "@/components/widget/onboarding/restart-setup";
 import { RotateCcw } from "lucide-react";
 import { useState } from "react";
-import { toast } from "sonner";
 
 interface Props {
   title: string;
-}
-
-/** Rewinds onboarding to the config-dir step. The backend clears the durable
- * facts the step machine derives progress from (and deletes a config dir or
- * parked import it materialized itself), then the preferences-changed event
- * re-routes the flow — no navigation needed here. */
-async function restartSetup() {
-  try {
-    await client.onboarding.reset();
-  } catch (error) {
-    // Typically the backend refusing to reset while a build is running.
-    toast.error(error instanceof Error ? error.message : String(error));
-    return;
-  }
-  onboardingActions.reset();
-  getTelemetry().captureEvent({ name: "onboarding_restarted" });
 }
 
 export function OnboardingHeader({ title }: Props) {
@@ -52,16 +32,7 @@ export function OnboardingHeader({ title }: Props) {
         Restart setup
       </button>
 
-      <ConfirmationDialog
-        open={confirming}
-        onOpenChange={setConfirming}
-        title="Restart setup from the beginning?"
-        message="Progress so far is discarded, and a configuration that was imported or created during setup is deleted. A pre-existing configuration directory you selected yourself is kept."
-        color="amber"
-        onConfirm={() => {
-          void restartSetup();
-        }}
-      />
+      <RestartSetupConfirmation open={confirming} onOpenChange={setConfirming} />
     </header>
   );
 }
